let SST1MContainer be field of DataContainer
