+++ ctapipe/io/containers.py
@@ -31,6 +31,45 @@ __all__ = ['InstrumentContainer',
            'HillasParametersContainer']
 
 
+class SST1MCameraContainer(Container):
+    pixel_flags = Field(None, 'numpy array containing pixel flags')
+    digicam_baseline = Field(None, 'Baseline computed by DigiCam')
+    local_camera_clock = Field(float, "camera timestamp")
+    gps_time = Field(float, "gps timestamp")
+    camera_event_type = Field(int, "camera event type")
+    array_event_type = Field(int, "array event type")
+    trigger_input_traces = Field(None, "trigger patch trace (n_patches)")
+    trigger_output_patch7 = Field(
+        None,
+        "trigger 7 patch cluster trace (n_clusters)")
+    trigger_output_patch19 = Field(
+        None,
+        "trigger 19 patch cluster trace (n_clusters)")
+
+    def fill_from_zfile_event(self, event):
+        self.pixel_flags = event.pixel_flags
+        self.digicam_baseline = event.baseline
+        self.local_camera_clock = event.local_time
+        self.gps_time = event.central_event_gps_time
+        self.camera_event_type = event.camera_event_type
+        self.array_event_type = event.array_event_type
+        self.trigger_input_traces = event.trigger_input_traces
+        self.trigger_output_patch7 = event.trigger_output_patch7
+        self.trigger_output_patch19 = event.trigger_output_patch19
+
+
+class SST1MContainer(Container):
+    tels_with_data = Field([], "list of telescopes with data")
+    tel = Field(
+        Map(SST1MCameraContainer),
+        "map of tel_id to SST1MCameraContainer")
+
+    def fill_from_zfile_event(self, event):
+        self.tels_with_data = [event.telescope_id, ]
+        for tel_id in self.tels_with_data:
+            self.tel[tel_id].fill_from_zfile_event(event)
+
+
 # todo: change some of these Maps to be just 3D NDarrays?
 
 
@@ -356,6 +395,7 @@ class DataContainer(Container):
     inst = Field(InstrumentContainer(), "instrumental information (deprecated")
     pointing = Field(Map(TelescopePointingContainer),
                      'Telescope pointing positions')
+    sst1m = Field(SST1MContainer(), "optional SST1M Specific Information")
 
 
 class MuonRingParameter(Container):
@@ -478,4 +518,3 @@ class HillasParametersContainer(Container):
 
     skewness = Field(0.0, 'measure of the asymmetry')
     kurtosis = Field(0.0, 'measure of the tailedness')
-
